+++ examples/lux_core_engine/LuxCoreEngineGeometry.cc
@@ -1,0 +1,4 @@
+#include "LuxCoreEngineGeometry.hh"
+
+using namespace ignition;
+using namespace rendering;
