+++ examples/lux_core_engine/LuxCoreEngineMaterial.cc
@@ -1,0 +1,8 @@
+#include "LuxCoreEngineMaterial.hh"
+
+using namespace ignition;
+using namespace rendering;
+
+LuxCoreEngineMaterial::LuxCoreEngineMaterial() {}
+
+LuxCoreEngineMaterial::~LuxCoreEngineMaterial() {}
