+++ examples/lux_core_engine/LuxCoreEngineObject.cc
@@ -1,0 +1,13 @@
+#include "LuxCoreEngineObject.hh"
+
+using namespace ignition;
+using namespace rendering;
+
+LuxCoreEngineObject::LuxCoreEngineObject() {}
+
+LuxCoreEngineObject::~LuxCoreEngineObject() {}
+  
+ScenePtr LuxCoreEngineObject::Scene() const
+{
+  return nullptr;
+}
