+++ examples/lux_core_engine/LuxCoreEngineSensor.cc
@@ -1,0 +1,8 @@
+#include "LuxCoreEngineSensor.hh"
+
+using namespace ignition;
+using namespace rendering;
+
+LuxCoreEngineSensor::LuxCoreEngineSensor() {}
+
+LuxCoreEngineSensor::~LuxCoreEngineSensor() {}
